+++ app/components/avo/views/resources/active_record/index.html.erb
