+++ app/components/avo/views/resources/active_record/show.html.erb
