+++ app/components/avo/views/resources/http/edit.html.erb
@@ -1,0 +1,49 @@
+<%= content_tag :div,
+  id: helpers.frame_id(@resource),
+  data: {
+    model_name: @resource.model_name.to_s,
+    resource_name: @resource.class.to_s,
+    record_id: @resource.record.id,
+    selected_resources_name: @resource.model_key,
+    selected_resources: [@resource.record.id],
+    **@resource.stimulus_data_attributes
+  } do %>
+  <%= render_cards_component %>
+  <%= form_with model: @resource.record,
+    scope: @resource.form_scope,
+    url: form_url,
+    method: form_method,
+    local: true,
+    html: {
+      novalidate: true
+    },
+    multipart: true do |form| %>
+    <%= render Avo::ReferrerParamsComponent.new back_path: back_path %>
+    <%= content_tag :div, class: 'space-y-12' do %>
+      <% @resource.get_items.each_with_index do |item, index| %>
+        <%= render Avo::Items::SwitcherComponent.new(
+          resource: @resource,
+          reflection: @reflection,
+          item: item,
+          index: index + 1,
+          view: @view,
+          parent_resource: @parent_resource,
+          parent_record: @parent_record,
+          form: form,
+          parent_component: self,
+          actions: @actions
+        )%>
+      <% end %>
+
+      <% if Avo.configuration.buttons_on_form_footers %>
+        <%= render Avo::PanelComponent.new do |c| %>
+          <% c.with_footer_tools do %>
+            <% @resource.render_edit_controls.each do |control| %>
+              <%= render_control control %>
+            <% end %>
+          <% end %>
+        <% end %>
+      <% end %>
+    <% end %>
+  <% end %>
+<% end %>
