+++ app/controllers/avo/active_record_resources_controller.rb
@@ -1,0 +1,6 @@
+require_dependency "avo/active_record_controller"
+
+module Avo
+  class ActiveRecordResourcesController < ActiveRecordController
+  end
+end
