+++ app/controllers/avo/application_controller.rb
@@ -242,7 +242,7 @@ module Avo
 
       if Avo.configuration.full_width_container
         contain = false
-      elsif Avo.configuration.full_width_index_view && action_name.to_sym == :index && self.class.superclass.to_s == "Avo::ResourcesController"
+      elsif Avo.configuration.full_width_index_view && action_name.to_sym == :index && self.class.superclass.to_s == "Avo::ActiveRecordResourcesController"
         contain = false
       end
