+++ app/controllers/avo/http_controller.rb
@@ -1,0 +1,7 @@
+module Avo
+  class HttpController < BaseController
+    def index
+      set_component_for __method__
+    end
+  end
+end
