+++ app/controllers/avo/http_resources_controller.rb
@@ -1,0 +1,6 @@
+require_dependency "avo/http_controller"
+
+module Avo
+  class HttpResourcesController < HttpController
+  end
+end
