--- app/controllers/avo/resources_controller.rb
@@ -1,6 +1,0 @@
-require_dependency "avo/base_controller"
-
-module Avo
-  class ResourcesController < BaseController
-  end
-end
