+++ spec/dummy/app/controllers/avo/active_record_attachments_controller.rb
@@ -1,4 +1,4 @@
 # This controller has been generated to enable Rails' resource routes.
 # More information on https://docs.avohq.io/2.0/controllers.html
-class Avo::ActiveRecordAttachmentsController < Avo::ResourcesController
+class Avo::ActiveRecordAttachmentsController < Avo::ActiveRecordResourcesController
 end
