+++ spec/dummy/app/controllers/avo/books_controller.rb
@@ -1,0 +1,2 @@
+class Avo::BooksController < Avo::HttpResourcesController
+end
