+++ spec/dummy/app/controllers/avo/cities_controller.rb
@@ -1,4 +1,4 @@
 # This controller has been generated to enable Rails' resource routes.
 # You shouldn't need to modify it in order to use Avo.
-class Avo::CitiesController < Avo::ResourcesController
+class Avo::CitiesController < Avo::ActiveRecordResourcesController
 end
