+++ spec/dummy/app/controllers/avo/compact_users_controller.rb
@@ -1,4 +1,4 @@
 # This controller has been generated to enable Rails' resource routes.
 # More information on https://docs.avohq.io/2.0/controllers.html
-class Avo::CompactUsersController < Avo::ResourcesController
+class Avo::CompactUsersController < Avo::ActiveRecordResourcesController
 end
