+++ spec/dummy/app/controllers/avo/course_links_controller.rb
@@ -1,2 +1,2 @@
-class Avo::CourseLinksController < Avo::ResourcesController
+class Avo::CourseLinksController < Avo::ActiveRecordResourcesController
 end
