+++ spec/dummy/app/controllers/avo/memberships_controller.rb
@@ -1,2 +1,2 @@
-class Avo::MembershipsController < Avo::ResourcesController
+class Avo::MembershipsController < Avo::ActiveRecordResourcesController
 end
