+++ spec/dummy/app/controllers/avo/products_controller.rb
@@ -1,4 +1,4 @@
 # This controller has been generated to enable Rails' resource routes.
 # You shouldn't need to modify it in order to use Avo.
-class Avo::ProductsController < Avo::ResourcesController
+class Avo::ProductsController < Avo::ActiveRecordResourcesController
 end
