+++ spec/dummy/app/controllers/avo/reviews_controller.rb
@@ -1,2 +1,2 @@
-class Avo::ReviewsController < Avo::ResourcesController
+class Avo::ReviewsController < Avo::ActiveRecordResourcesController
 end
