+++ spec/dummy/app/controllers/avo/spouses_controller.rb
@@ -1,2 +1,2 @@
-class Avo::SpousesController < Avo::ResourcesController
+class Avo::SpousesController < Avo::ActiveRecordResourcesController
 end
