+++ spec/dummy/app/controllers/avo/team_users_controller.rb
@@ -1,4 +1,4 @@
 # This controller has been generated to enable Rails' resource routes.
 # You shouldn't need to modify it in order to use Avo.
-class Avo::TeamUsersController < Avo::ResourcesController
+class Avo::TeamUsersController < Avo::ActiveRecordResourcesController
 end
